test: pass appropriate error handler for antlr adapter
