+++ server/engine/src/test/java/org/eclipse/lsp/cobol/usecases/TestErrorOnReadyTraceOutsideProcedureDivision.java
@@ -23,7 +23,6 @@ import org.eclipse.lsp.cobol.test.engine.UseCaseEngine;
 import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.DiagnosticSeverity;
 import org.eclipse.lsp4j.Range;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 // to review
 
@@ -54,7 +53,6 @@ class TestErrorOnReadyTraceOutsideProcedureDivision {
   }
 
   @Test
-  @Disabled("EXPERIMENTAL_COBOL")
   void testHw() {
     UseCaseEngine.runTest(
         TEXT,
