+++ server/engine/src/test/java/org/eclipse/lsp/cobol/usecases/TestHwUnexpectedTextInIdentificationDivision.java
@@ -23,7 +23,6 @@ import org.eclipse.lsp.cobol.test.engine.UseCaseEngine;
 import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.DiagnosticSeverity;
 import org.eclipse.lsp4j.Range;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 /** This test checks that unexpected text in ID DIVISION is flagged */
@@ -40,7 +39,6 @@ class TestHwUnexpectedTextInIdentificationDivision {
           + "           DISPLAY {$VARNAME}.";
 
   @Test
-  @Disabled("EXPERIMENTAL_COBOL")
   void test() {
     UseCaseEngine.runTest(
         TEXT,
