+++ server/engine/src/test/java/org/eclipse/lsp/cobol/usecases/TestNoNullPointerExpForIncompleteReplaceStatement.java
@@ -23,7 +23,6 @@ import org.eclipse.lsp.cobol.test.engine.UseCaseEngine;
 import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.DiagnosticSeverity;
 import org.eclipse.lsp4j.Range;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -63,7 +62,6 @@ class TestNoNullPointerExpForIncompleteReplaceStatement {
   }
 
   @Test
-  @Disabled("EXPERIMENTAL_COBOL")
   void testHwWhenTypingEndsAtDoubleEqualChar() {
     UseCaseEngine.runTest(
         TEXT,
@@ -92,7 +90,6 @@ class TestNoNullPointerExpForIncompleteReplaceStatement {
   }
 
   @Test
-  @Disabled("EXPERIMENTAL_COBOL")
   void testHwWhenTypingEndsAtSingleEqualChar() {
     UseCaseEngine.runTest(
         TEXT2,
