+++ server/engine/src/test/java/org/eclipse/lsp/cobol/usecases/TestSqlIncludeStatementNotDefinedCorrectly.java
@@ -24,7 +24,6 @@ import org.eclipse.lsp.cobol.common.error.ErrorSource;
 import org.eclipse.lsp.cobol.test.engine.UseCaseEngine;
 import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.Range;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 /** This test checks sql include statement if it is defined correctly. */
@@ -52,7 +51,6 @@ class TestSqlIncludeStatementNotDefinedCorrectly {
   }
 
   @Test
-  @Disabled("EXPERIMENTAL_COBOL")
   void testHw() {
     UseCaseEngine.runTest(
         TEXT,
