+++ server/engine/src/test/java/org/eclipse/lsp/cobol/usecases/example/TestParametrised.java
@@ -16,18 +16,16 @@ package org.eclipse.lsp.cobol.usecases.example;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import java.util.stream.Stream;
 import org.eclipse.lsp.cobol.common.dialects.CobolLanguageId;
 import org.eclipse.lsp.cobol.common.error.ErrorSource;
 import org.eclipse.lsp.cobol.test.engine.UseCaseEngine;
 import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.DiagnosticSeverity;
 import org.eclipse.lsp4j.Range;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-
-import java.util.stream.Stream;
 
 /** UseCase test example with variants */
 class TestParametrised {
@@ -65,7 +63,6 @@ class TestParametrised {
   @ParameterizedTest
   @MethodSource("textsGetter")
   @DisplayName("Parameterized - different ends")
-  @Disabled("EXPERIMENTAL_COBOL")
   void testHw(String text) {
     UseCaseEngine.runTest(
         text,
