+++ server/engine/src/test/java/org/eclipse/lsp/cobol/usecases/example/TestWithErrorCheck.java
@@ -22,7 +22,6 @@ import org.eclipse.lsp.cobol.test.engine.UseCaseEngine;
 import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.DiagnosticSeverity;
 import org.eclipse.lsp4j.Range;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 /** UseCase test example when you expecting errors */
@@ -51,7 +50,6 @@ class TestWithErrorCheck {
   }
 
   @Test
-  @Disabled("EXPERIMENTAL_COBOL")
   void testHw() {
     UseCaseEngine.runTest(
         TEXT,
